Show memory full text only for internal storage
When the internal memory is full, a warning text is shown
in Settings/Storage. This warning was displayed for all
types of memory (internal/primary, sd card and usb storage)
even though it was only referring to the primary memory.
With this fix, it is only shown for the primary memory.

Change-Id: I35ad4b0b45c6752ddc9d7166495435b27fd4a170