Settings: Recents RAM bar (2/2)
due that the old rambar was resource consuming and as the settings rambar not really
right in showing the truth of real RAM usage in android, I decided to go new ways.

This commit changes our rambar (or better add a total new one) in following points:

- we do not use any longer ther permanent observed RunningState machine
- instead we read out /proc/meminfo (thanks steven for that idea)
- rambar shows now the real memory usage
- ppl have the possibility to show applications, cache and other ram usage as they like
- deeper explanations I add how android mem managment works.
- color picker for all rambar states
- reset to default
- updates only on actions on recents so that it is making sense
- and a lot other

cheers

Many thanks to @xLaMbChOpSx for forward porting this from SlimBean.

Change-Id: I6f79c253f6884dffc8bcb5e1b52421143f91e02d