Fix Settings app crash for devices with no USB support.
On emulator and any device that does not support USB connection, the
Settings app crashes as soon as the user tries to view the details of
an installed app. This is because the InstalledAppDetails activity
assumes that the USB service is always started, which is no longer
true due to a change in SystemServer.java (frameworks/base, Change-Id
I6dad511c652bcacc085f27ede5f8f22a6982c6bd). Adding null checks fixes
this issue.

Change-Id: I85263fb795ee06bd61ba28ec12a7161c1f17f4cc
Signed-off-by: Yu Ning <yu.ning@intel.com>